% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_geodata.R
\name{as_geodata}
\alias{as_geodata}
\title{Promote a data frame to an \code{sf} object}
\usage{
as_geodata(input_data, coords, crs, ..., verbose = TRUE)
}
\arguments{
\item{input_data}{data frame or tibble}

\item{coords}{names of columns; see \code{\link[sf:st_as_sf]{sf::st_as_sf()}}}

\item{crs}{can use integer or \code{\link[sf:st_crs]{sf::st_crs()}}}

\item{...}{further arguments to \code{\link[sf:st_as_sf]{sf::st_as_sf()}}}

\item{verbose}{logical}
}
\description{
Promote a data frame to an \code{sf} object
}
