% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_gridded.R
\name{as_gridded}
\alias{as_gridded}
\title{as_gridded}
\usage{
as_gridded(x, layer, coord_vars = NULL, crs = NULL)
}
\arguments{
\item{x}{\code{sf} or \verb{Spatial*} object}

\item{layer}{name of layer to grid}

\item{coord_vars}{names of coordinate variables}

\item{crs}{suitable for use with \code{proj4string(.) <- crs}}
}
\description{
as_gridded
}
