% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_raster.R
\name{as_raster}
\alias{as_raster}
\title{as_raster}
\usage{
as_raster(x, layer, coord_vars = NULL, crs = NULL, ...)
}
\arguments{
\item{x}{passed to \code{as_gridded()}}

\item{layer}{layer name}

\item{coord_vars}{names of coordinates (optional)}

\item{crs}{passed to \code{as_gridded()}}

\item{...}{passes to \code{raster::raster()}}
}
\description{
Rasterize geodata
}
