% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meters.R
\name{as_square_meters}
\alias{as_square_meters}
\title{as_square_meters}
\usage{
as_square_meters(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Convert a plain numeric vector to a \code{units} object with units of \code{m^2}
}
