% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_rasters.R
\name{bind_rasters}
\alias{bind_rasters}
\title{bind_rasters}
\usage{
bind_rasters(
  x,
  ...,
  fill = NA,
  outfile = NULL,
  verbose = getOption("verbose", default = FALSE)
)
}
\arguments{
\item{x}{RasterLayer}

\item{...}{one or more RasterLayers}

\item{fill}{ignored}

\item{outfile}{character}

\item{verbose}{logical}
}
\value{

}
\description{
bind_rasters
}
