% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{build.SRTM.terrain.90m.ras.fun}
\alias{build.SRTM.terrain.90m.ras.fun}
\title{Build a high resolution topographic rasters stack (elevation, slope, aspect)}
\usage{
build.SRTM.terrain.90m.ras.fun(country_code.chr, NAME_1.chr = NULL,
  aggregation_factor.num = NULL, EPSG.chr = NULL, path.chr)
}
\arguments{
\item{country_code.chr}{a character specifying the ISO contrycode. Ex : BE for belgium}

\item{NAME_1.chr}{a character specifying the NAME_1 value for lower than country level information}

\item{aggregation_factor.num}{a numeric specifying the aggregation factor to get the desired spatial resolution}

\item{EPSG.chr}{a character specifying the EPSG code of the desired Coordiante Reference System (CRS)}

\item{path.chr}{a character specifying the path where to dowload the SRTM data}
}
\value{
A stack of topographic rasters
}
\description{
This function builds a topographic rasters stack for Belgium (elevation, slope, aspect). The rasters are projected in the
3812 EPSG code.
}
\author{
Thomas Goossens - pokyah.github.io
}
