% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{build.static.ggmap}
\alias{build.static.ggmap}
\title{Build a static map displaying predictions and their related error}
\usage{
build.static.ggmap(gridded.data.df, boundaries.sf, layer.error.bool = FALSE,
  legend.error.bool = FALSE, pretty_breaks.bool = TRUE, title.chr,
  legend.chr, target.chr, nb_classes.num, reverse_pal.bool = TRUE,
  resolution.chr = NULL)
}
\arguments{
\item{gridded.data.df}{A data frame obtained from a SpatialGridDataFrame containing data}

\item{boundaries.sf}{A sf containing data from Wallonia boundaries}

\item{layer.error.bool}{A boolean specifying if you want to display the layer with error}

\item{legend.error.bool}{A boolean specifying if you want to display the legend of the error layer}

\item{pretty_breaks.bool}{A boolean specifying the type of legend you want. TRUE for pretty breaks, FALSE for quantile scale}

\item{title.chr}{A character specifying the title you want for your map}

\item{target.chr}{A character specifying the name of the column containing the predicted parameter.}

\item{nb_classes.num}{A numeric for the number of classes you want}

\item{reverse_pal.bool}{A boolean if you want to reverse color palette. Default is TRUE, which means that 'red' is for high values and 'blue' for low values}

\item{resolution.chr}{A character which specifies the resolution of the map}

\item{target.name}{the type of weather param : tsa, plu, ind_plu, etc...}
}
\value{
a ggplot map object
}
\description{
Build a static map displaying predictions and their related error
}
\author{
Loïc Davadan <- ldavadan.github.io
}
