% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{build.vs.grid.fun}
\alias{build.vs.grid.fun}
\title{Build a sp/sf interpolation grid with the desired spatial resolution for Wallonia}
\usage{
build.vs.grid.fun(country_code.chr, NAME_1.chr, res.num, geom.chr, sf.bool,
  EPSG.chr = NULL)
}
\arguments{
\item{country_code.chr}{a character specifying the ISO contrycode. Ex : BE for belgium}

\item{NAME_1.chr}{a character specifying the NAME_1 value for lower than country level information}

\item{res.num}{A numeric representing the spatial resolution of the desired grid (in meters)}

\item{geom.chr}{A character specifying the geometry of the interpolation grid. Cant take a value among ("polygons", "centers" or "corners")}

\item{sf.bool}{A boolean specifying if we want as sp or sf (TRUE for sf)}

\item{EPSG.chr}{A character specifying the EPSG code of the desired Coordiante Reference System (CRS)}
}
\value{
A sf spatial grid with the desired resolution clipped to the Wallonia Polygon
}
\description{
Inspired from https://stackoverflow.com/questions/41787313/how-to-create-a-grid-of-spatial-points,
https://www.nceas.ucsb.edu/~frazier/RSpatialGuides/OverviewCoordinateReferenceSystems.pdf, https://gis.stackexchange.com/questions/22843/converting-decimal-degrees-units-to-km-in-r,
https://stackoverflow.com/questions/48727511/r-grid-of-points-from-polygon-input
}
\author{
Thomas Goossens - pokyah.github.io
}
