% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{build_leaflet_template.fun}
\alias{build_leaflet_template.fun}
\title{Build a responsive leaflet map displaying agromet AWS network data}
\usage{
build_leaflet_template.fun(records.sf)
}
\arguments{
\item{records.sf}{A sf containing the records to be displayed}
}
\value{
a leaflet map object
}
\description{
Build a responsive leaflet map displaying agromet AWS network data
}
\author{
Thomas Goossens - pokyah.github.io
}
