% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{build_lowRes_terrain_rasters.fun}
\alias{build_lowRes_terrain_rasters.fun}
\title{Build a topographic rasters stack (elevation, slope, aspect)}
\usage{
build_lowRes_terrain_rasters.fun()
}
\value{
A stack of topographic rasters
}
\description{
This function builds a topographic rasters stack for Belgium (elevation, slope, aspect). The rasters are projected in the
3812 EPSG code. No input parameters are required.
}
\author{
Thomas Goossens - pokyah.github.io
}
