% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotools-package.R
\docType{data}
\name{CA_ALBERS_NAD83}
\alias{CA_ALBERS_NAD83}
\title{California Albers (EPSG:3310)}
\format{
an \code{sp::CRS} object
}
\usage{
CA_ALBERS_NAD83
}
\description{
California Albers (EPSG 3310)
}
\keyword{datasets}
