% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_data.R
\name{coord_data}
\alias{coord_data}
\title{Return the coordinates of a Spatial* object as a data.frame}
\usage{
coord_data(spobj)
}
\description{
Return the coordinates of a Spatial* object as a data.frame
}
