% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_web_mercator.R
\name{coord_web_mercator}
\alias{coord_web_mercator}
\title{coord_web_mercator}
\usage{
coord_web_mercator(envelope = NULL, xlim = NULL, ylim = NULL)
}
\arguments{
\item{envelope}{sf or sfc}

\item{xlim}{numeric (two values), if \code{envelope} isn't provided}

\item{ylim}{numeric (two values), if \code{envelope} isn't provided}
}
\value{
output of \code{coord_sf(...)}
}
\description{
coord_web_mercator
}
