% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates_to_columns.R
\name{coordinates_to_columns}
\alias{coordinates_to_columns}
\title{Copy coordinates into (two) columns}
\usage{
coordinates_to_columns(geodata, coords, ...,
  verbose = getOption("verbose"))
}
\arguments{
\item{geodata}{(sf or sfc)}

\item{coords}{(character) names of columns}

\item{...}{further arguments}

\item{verbose}{(logical)}
}
\description{
Copy coordinates into (two) columns
}
