% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_raster.R
\name{cut_raster}
\alias{cut_raster}
\title{cut_raster}
\usage{
cut_raster(
  raster_layer,
  breaks,
  include.lowest = TRUE,
  ...,
  na.rm = TRUE,
  verbose = getOption("verbose", default = FALSE)
)
}
\arguments{
\item{raster_layer}{\href{raster::Raster-class}{RasterLayer} object}

\item{breaks}{passed to \code{\link[=cut]{cut()}}}

\item{include.lowest}{passed to \code{\link[=cut]{cut()}}}

\item{...}{passed to \code{\link[=cut]{cut()}}}

\item{na.rm}{logical; passed to \code{\link[=cut]{cut()}}}

\item{verbose}{logical}
}
\value{
a new \href{raster::Raster-class}{RasterLayer} object, of the same dimensions, etc., with values cut accordingly
}
\description{
cut_raster
}
