% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/drop_geometry.R
\name{drop_geometry}
\alias{drop_geometry}
\title{Coerce an \code{sf} object to a \code{tibble}}
\usage{
drop_geometry(geodata)

drop_geometry(geodata)
}
\arguments{
\item{geodata}{an \code{sf} object}
}
\value{
a \code{tibble}

a \code{tibble}
}
\description{
Coerce an \code{sf} object to a \code{tibble}

Coerce an \code{sf} object to a \code{tibble}
}
