% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotools-package.R
\docType{data}
\name{EPSG_4326}
\alias{EPSG_4326}
\alias{WGS84_GPS}
\title{EPSG 4326}
\format{
an \code{sp::CRS} object

integer (EPSG code)
}
\usage{
EPSG_4326

WGS84_GPS
}
\description{
WGS84 (EPSG 4326)

WGS84 (EPSG 4326)
}
\keyword{datasets}
