% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_layer.R
\name{extract_layer}
\alias{extract_layer}
\title{To extract a layer from an (unzipped) shapefile}
\usage{
extract_layer(dsn, layer, datum, quiet = FALSE, ...)
}
\arguments{
\item{dsn}{data source name}

\item{layer}{layer name}

\item{datum}{(optional) passed to \code{\link[sp]{spTransform}}}

\item{quiet}{(optional)}

\item{\dots}{further arguments passed to \code{\link[rgdal]{readOGR}}}
}
\description{
To extract a layer from an (unzipped) shapefile
}
\examples{
\dontrun{

basins <- extract_layer(
  unzip_url("http://www.arb.ca.gov/ei/gislib/boundaries/ca_air_basins.zip",
            exdir = file.path("cache")),
  layer = "CaAirBasin")

districts <- extract_layer(
  unzip_url("http://www.arb.ca.gov/ei/gislib/boundaries/ca_air_district.zip",
            exdir = file.path("cache")),
  layer = "CaAirDistrict")

counties <- extract_layer(
  unzip_url("http://www.arb.ca.gov/ei/gislib/boundaries/ca_county.zip",
            exdir = file.path("cache")),
  layer = "CaCounty")
}

}
