% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_ids.R
\name{feature_ids}
\alias{feature_ids}
\alias{feature_ids.SpatialPolygons}
\alias{feature_ids.SpatialPolygonsDataFrame}
\title{To get the IDs of features in a SpatialPolygons* object}
\usage{
feature_ids(spobj, ...)

\method{feature_ids}{SpatialPolygons}(spobj, ...)

\method{feature_ids}{SpatialPolygonsDataFrame}(spobj, ...)
}
\arguments{
\item{spobj}{SpatialPolygons* object}

\item{\dots}{(unused)}
}
\description{
To get the IDs of features in a SpatialPolygons* object
}
