% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-filter.R
\name{filter_.SpatialDataFrame}
\alias{filter_.SpatialDataFrame}
\title{Filter (subset) a Spatial*DataFrame object}
\usage{
\method{filter_}{SpatialDataFrame}(object, ..., .dots)
}
\arguments{
\item{object}{a \code{Spatial*DataFrame}}

\item{...}{see \link{subset}}

\item{.dots}{}
}
\value{
a subset of the original \code{Spatial*DataFrame}
}
\description{
Filter (subset) a Spatial*DataFrame object
}
\examples{
library(dplyr)
library(sp)
data(meuse, package = "sp")
coordinates(meuse) <- ~x+y
df <- filter(meuse, copper > median(copper))
stopifnot(inherits(df, class(meuse)))

}
\seealso{
subset
}
