% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_spatial.R
\name{filter_spatial}
\alias{filter_spatial}
\title{Apply a spatial filter}
\usage{
filter_spatial(
  x,
  y,
  FUN = NULL,
  crs = NULL,
  ...,
  verbose = getOption("verbose")
)
}
\arguments{
\item{x}{(\link[sp:Spatial-class]{Spatial} object) to be filtered}

\item{y}{(\link[sp:Spatial-class]{Spatial} object) to filter with}

\item{FUN}{(function) a spatial predicate, for example: \link[rgeos:pred-binary-gIntersects]{gIntersects}}

\item{...}{further arguments to \code{FUN}}

\item{verbose}{(logical)}
}
\value{
(typically) a subset of \code{spobj1}
}
\description{
Apply a spatial filter
}
