% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{force_crs}
\alias{force_crs}
\title{Force replacement of CRS while leaving coordinate values unchanged}
\usage{
force_crs(geodata, crs, ..., verbose = getOption("verbose"))
}
\arguments{
\item{geodata}{\code{sf} or \code{sfc} object}

\item{crs}{passed to `sf::st_as_sf()``}

\item{...}{further arguments to \code{\link[sf:st_as_sf]{sf::st_as_sf()}}}

\item{verbose}{display messages}
}
\description{
Force replacement of CRS while leaving coordinate values unchanged
}
