% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_xy.R
\name{fortify_xy}
\alias{fortify_xy}
\title{fortify_xy}
\usage{
fortify_xy(spobj)
}
\arguments{
\item{spobj}{\verb{Spatial*} object}
}
\description{
For use with ggplot2 (plotting polygons)
}
\note{
Consider using \code{ggplot2::geom_sf()} instead.
}
