\name{fr_geo_deg}
\alias{fr_geo_deg}
\alias{fr_geo_rad}
\title{Dataset french zip code}
\usage{
data(fr_geo_deg)
data(fr_geo_rad)
}
\description{French dataset for zip code (postal code).

Data set, from French postal data can be retreive here:
\url{ http://www.galichon.com/codesgeo/}.


Other data are here, \url{http://earth-info.nga.mil/gns/html/cntry_files.html}.

}

\keyword{datasets}
