% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gClip.R
\name{gClip}
\alias{gClip}
\title{Clip one SpatialPolygons* object with another SpatialPolygons* object}
\usage{
gClip(spobj1, spobj2, ...)
}
\description{
Clip one SpatialPolygons* object with another SpatialPolygons* object
}
\note{
You're responsible for proper handling of numeric attributes.
}
