% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{gContainsOrOverlaps}
\alias{gContainsOrOverlaps}
\title{gContainsOrOverlaps}
\usage{
gContainsOrOverlaps(geom2, geom1, byid = TRUE)
}
\arguments{
\item{geom2}{\verb{Spatial*} object}

\item{geom1}{\verb{Spatial*} object}

\item{byid}{passed to \code{rgeos::gContains()} and \code{rgeos::gOverlaps()}}
}
\description{
This is the default predicate used by \code{gFilter()}.
}
\note{
Unlike most \code{rgeos} functions, this returns row.names !!
}
