% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode}
\alias{geocode}
\title{Geocode street address(es)}
\usage{
geocode(address, ..., key, verbose = FALSE)
}
\arguments{
\item{address}{character}

\item{\dots}{further arguments to be passed on to the Google Maps API}

\item{key}{character (your valid API key)}

\item{verbose}{TRUE or FALSE}
}
\description{
Geocode street address(es) using the Google Maps API.
}
\note{
\code{GOOGLE_API_KEY} should be set in \code{~/.Renviron} or \code{/etc/R/Renviron.site}
}
\references{
https://developers.google.com/maps/documentation/geocoding/
}
