% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomatching.R
\name{geomatching}
\alias{geomatching}
\title{Align heterogeneous space-time data onto a common grid}
\usage{
geomatching(data, settings = NULL, aggregate = FALSE, group_by = "mun",
sd = "intra", stats = "mean")
}
\arguments{
\item{data}{List of space-time datasets, each either a \code{data.frame} or a
3D \code{matrix}. The spatial grid of the first element is used as the
reference grid for aligning all others.}

\item{settings}{Named list with the fields \strong{format}, \strong{type}
and \strong{crs}. \code{"format"} defines the data format for each input:
use \code{"xyt"} if the i-th input is a data.frame with three columns
representing x-coordinate, y-coordinate and time; use \code{"matrix"} if the
i-th input is a 3D array where the first, second, and third dimensions
correspond to x, y, and time respectively. \code{"type"} specifies the
geometry type used for matching, e.g., \code{"points"} or \code{"grid"}.
\code{"crs"} sets the Coordinate Reference System as an EPSG code.}

\item{aggregate}{Logical flag. If \code{TRUE}, the function aggregates the
matched data over the specified administrative boundaries (see group_by)
and computes summary statistics (e.g., mean, median, quartiles, standard
deviation) for each space-time variable. If \code{FALSE}, the function returns
the matched data without aggregation.}

\item{group_by}{Character string. Specifies the administrative boundary level
for aggregation when \code{aggregate = TRUE}. Accepted values are \strong{"mun"}
(municipality), \strong{"prov"} (province) or \strong{"reg"} (region).
The data will be grouped and summarized according to the selected boundary level.}

\item{sd}{Character string. Specifies the type of uncertainty desired.
Accepted values are \strong{"intra"} or \strong{"varX"}. Use \strong{"intra"}
if the standard deviation is calculated as the intra-variability within the polygon.
Use \strong{"varX"} if the variables has a standard deviation associated. In this case,
for each column harmonised, an other column called the same plus "_sd" is looked for.
For each polygon, these values are squared, summed and divided by the square of
elements used. Finally, square root is provided.}

\item{stats}{Use only if aggregate is \code{TRUE}. Select which statistics
you want to calculate. Accepted values are: \strong{min},\strong{1st_quartile},
\strong{mean},\strong{median},\strong{3rd_quartile},\strong{max},\strong{sd}.}
}
\value{
A \strong{data.frame} containing the matched data aligned on the
reference grid.
}
\description{
Aligns input space-time data from different spatial grids onto a specific
common grid, even when their geographic reference systems may differ.
This procedure is known as \strong{spatial overlay}.
}
\examples{
# SEE "demo.Rmd" FOR MORE DETAILS

\dontrun{
res_geomatch <- geomatching(
  data=list(AQ_EEA_NO2, AQ_CAMS_NO2),
  settings = list(
    "format"=list("xyt", "matrix"),
    "type"=list("points", "grid"),
    "crs"=list(4979, 4326)
  )
)}

}
\seealso{
\url{https://github.com/GRINS-Spoke0-WP2/geotools/blob/develop/demo/demo.Rmd}
}
