% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{get.points.cover_pct.fun}
\alias{get.points.cover_pct.fun}
\title{Get cover classes percentages for buffered points with custom radius}
\usage{
get.points.cover_pct.fun(cover.sf, points.sf, radius.num)
}
\arguments{
\item{cover.sf}{a sf polygon of the land cover}

\item{radius.num}{a numeric specifying the radius of the buffere th corine shapefiles resides}

\item{points}{a sf points of the locations on which surrounding cover needs to be summarized}
}
\value{
a sf containing the %
}
\description{
Get cover classes percentages for buffered points with custom radius
}
\author{
Thomas Goossens - pokyah.github.io
}
