% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_map_RSP.R
\name{ggplot_map_RSP}
\alias{ggplot_map_RSP}
\title{ggplot_map_RSP}
\usage{
ggplot_map_RSP(
  data = NULL,
  tileset = NULL,
  tile_opacity = NULL,
  coastline = NULL,
  zoom = NULL,
  ...,
  envelope = NULL,
  crs = WEB_MERCATOR,
  expand = ggplot2::expansion(mult = 0.025),
  verbose = getOption("verbose")
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
ggplot_map_RSP
}
