% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_tileset.R
\name{ggplot_tileset}
\alias{ggplot_tileset}
\title{ggplot_tileset}
\usage{
ggplot_tileset(
  url,
  envelope,
  zoom = NULL,
  expand = c(0, 0),
  max_tiles = NULL,
  verbose = getOption("verbose", default = FALSE)
)
}
\description{
ggplot_tileset
}
