% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{gSubset}
\alias{gSubset}
\title{gSubset}
\usage{
gSubset(spobj1, spobj2)
}
\arguments{
\item{spobj1}{\verb{Spatial*} object}

\item{spobj2}{\verb{SpatialPolygons*} object}
}
\description{
Subset a Spatial* object using a SpatialPolygons* object
}
