% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.R
\name{gtl_calculate_psummer}
\alias{gtl_calculate_psummer}
\title{Internal function to determine the summer precipitation months}
\usage{
gtl_calculate_psummer(prec, lat)
}
\arguments{
\item{prec}{A vector with precipitation for each month}

\item{lat}{The latitude}
}
\value{
A vector with the precipitation in the "summer" month as defined for Köppen climate classification
}
\description{
Internal function to determine the summer precipitation months
}
\keyword{internal}
