% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{gtl_chk_api_call}
\alias{gtl_chk_api_call}
\title{Check API connexion}
\usage{
gtl_chk_api_call(call)
}
\arguments{
\item{call}{An API url}
}
\value{
Returns an error if the API call fails
}
\description{
Check API connexion
}
