% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{gtl_chk_idb_api_key}
\alias{gtl_chk_idb_api_key}
\title{Check IDB API key}
\usage{
gtl_chk_idb_api_key()
}
\value{
Returns an error if the IDB API key is not properly set
}
\description{
Function to check if the IDB API key is correctly set.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# Not run: fails if the IDB API key isn't set

gtl_chk_idb_api_key()
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[idbr:idb_api_key]{idbr::idb_api_key()}}
}
