% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{gtl_country_list_wb}
\alias{gtl_country_list_wb}
\title{World Bank country list}
\usage{
gtl_country_list_wb()
}
\description{
This function returns a list of World Bank countries
for a choice selector in a Shiny app.
}
\examples{
gtl_country_list_wb()

}
