% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{gtl_crs_proj}
\alias{gtl_crs_proj}
\title{Get a defined CRS for a given code}
\usage{
gtl_crs_proj(code)
}
\arguments{
\item{country}{A string with the country name}
}
\value{
A crs object
}
\description{
Get a defined CRS for a given code
}
