% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{gtl_crs_regional}
\alias{gtl_crs_regional}
\title{Get a defined proj4 string for a given country}
\usage{
gtl_crs_regional(country)
}
\arguments{
\item{country}{A string with the country name}
}
\value{
A CRS object
}
\description{
A function that returns an ideal CRS for a given
country.
}
\details{
The current list of possible countries:
\itemize{
\item Russia
}
}
\examples{
gtl_crs_regional("Russia")

}
