% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{gtl_dwnl_api_json}
\alias{gtl_dwnl_api_json}
\title{Download JSON response from API}
\usage{
gtl_dwnl_api_json(call)
}
\arguments{
\item{call}{An API url}
}
\value{
JSON body
}
\description{
Download JSON response from API
}
\examples{
gtl_dwnl_api_json("https://dummyjson.com/products/1")

}
