% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{gtl_full_date}
\alias{gtl_full_date}
\title{Format full date}
\usage{
gtl_full_date(date)
}
\arguments{
\item{date}{A date in YYYY-MM-DD format}
}
\value{
A formated \verb{day month year} (\%d \%B \%Y) date
}
\description{
Format full date
}
