% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{gtl_gis_tissot_indicatrix}
\alias{gtl_gis_tissot_indicatrix}
\title{World Tissot Matrix}
\usage{
gtl_gis_tissot_indicatrix()
}
\value{
A simple feature layer
}
\description{
World Tissot Matrix
}
\examples{
gtl_gis_tissot_indicatrix()
}
