% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_tools.R
\name{gtl_hc_color_axis}
\alias{gtl_hc_color_axis}
\title{Return a list of color breaks for Higchart maps}
\usage{
gtl_hc_color_axis(breaks, palette)
}
\arguments{
\item{breaks}{The data with the breaks (using chopped data with \code{\link[santoku:santoku]{santoku::santoku}})}

\item{palette}{A color palette that can take a numeric argument}
}
\value{
A list of lists to use in colorAxis' dataClasses argument
}
\description{
Return a list of color breaks for Higchart maps
}
\examples{
data <- seq(1, 20)
data_cut <- santoku::chop(data, breaks = c(5, 10, 15))
gtl_hc_color_axis(data_cut, rainbow)
}
