% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_tools.R
\name{gtl_hc_discrete_color_list}
\alias{gtl_hc_discrete_color_list}
\title{Function to build the highchart colorAxis list}
\usage{
gtl_hc_discrete_color_list(...)
}
\arguments{
\item{...}{lvls and colors}
}
\description{
Function to build the highchart colorAxis list
}
\keyword{internal}
