% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_tools.R
\name{gtl_hc_mapView}
\alias{gtl_hc_mapView}
\title{mapView options for highcharter objects}
\usage{
gtl_hc_mapView(hc, ...)
}
\arguments{
\item{hc}{A \code{highchart} htmlwidget object.}

\item{...}{Options as defined in \url{https://api.highcharts.com/highmaps/mapView}.}
}
\description{
mapView options for highcharter objects
}
