% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{gtl_hmd_codes}
\alias{gtl_hmd_codes}
\title{Provide HMD country codes}
\usage{
gtl_hmd_codes()
}
\value{
A dataframe with 2 columns and 47 rows
}
\description{
This function returns a dataframe with the 47 countries for which
the \href{https://www.mortality.org}{Human Mortality Database} provides
data. For each country there is a \code{name} and a \code{code}.
}
\examples{
gtl_hmd_codes()

}
