% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.R
\name{gtl_koppen_a_climates}
\alias{gtl_koppen_a_climates}
\title{Internal function to determine A climates}
\usage{
gtl_koppen_a_climates(prec, ptot, pwinter, psummer)
}
\arguments{
\item{prec}{Precipitations}

\item{ptot}{Total precipitation}

\item{pwinter}{Winter precipitations}

\item{psummer}{Summer precipitations}
}
\value{
A string
}
\description{
Internal function to determine A climates
}
\keyword{internal}
