% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.R
\name{gtl_koppen_b_climates}
\alias{gtl_koppen_b_climates}
\title{Internal function to determine B climates}
\usage{
gtl_koppen_b_climates(tavg, ptot, e_thresh)
}
\arguments{
\item{tavg}{Average temperatures}

\item{ptot}{Total precipitation}

\item{e_thresh}{Evaporation threshold}
}
\value{
A string
}
\description{
Internal function to determine B climates
}
\keyword{internal}
