% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.R
\name{gtl_koppen_c_climates}
\alias{gtl_koppen_c_climates}
\title{Internal function to determine B climates}
\usage{
gtl_koppen_c_climates(temp, prec, pwinter, psummer)
}
\arguments{
\item{temp}{Temperatures}

\item{prec}{Precipitations}

\item{pwinter}{Winter precipitations}

\item{psummer}{Summer precipitations}
}
\value{
A string
}
\description{
Internal function to determine B climates
}
\keyword{internal}
