% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.R
\name{gtl_koppen_e_climates}
\alias{gtl_koppen_e_climates}
\title{Internal function to determine E climates}
\usage{
gtl_koppen_e_climates(temp)
}
\arguments{
\item{temp}{Temperatures}
}
\value{
A string
}
\description{
Internal function to determine E climates
}
\keyword{internal}
