% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{gtl_opt_long_language}
\alias{gtl_opt_long_language}
\title{Return long language name}
\usage{
gtl_opt_long_language()
}
\value{
A string with the language name (lowercase)
}
\description{
Function to get the current language setting in full.
}
\examples{
# With a french locale
gtl_opt_set_i18n("fr_CH")
gtl_opt_long_language()

# With an english locale
gtl_opt_set_i18n("en_US")
gtl_opt_long_language()
}
\seealso{
\code{\link[=gtl_opt_set_i18n]{gtl_opt_set_i18n()}}
}
