% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{gtl_opt_set_i18n}
\alias{gtl_opt_set_i18n}
\title{Set language and country settings}
\usage{
gtl_opt_set_i18n(code)
}
\arguments{
\item{code}{A valid i18n code (example: fr_CH)}
}
\description{
This function allows setting the default language and country
settings using a valid i18n code.
}
\examples{
# Swiss french locale
gtl_opt_set_i18n("fr_CH")
gtl_options("language")
gtl_options("country")

# American english locale
gtl_opt_set_i18n("en_US")
gtl_options("language")
gtl_options("country")
}
