% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{gtl_opt_short_language}
\alias{gtl_opt_short_language}
\title{Return short language code}
\usage{
gtl_opt_short_language(valid = c("en", "fr"))
}
\arguments{
\item{valid}{A vector with valid language options}
}
\value{
A string
}
\description{
Function that returns the short language code currently set
if it is in the list of \emph{valid} options. If the current language
isn't in the list of \emph{valid} options, the first element of the
valid elements is returned.
}
\examples{
gtl_opt_set_i18n("fr_CH")
gtl_opt_short_language(valid = c("de", "fr"))

# With the current local not among the valid options
gtl_opt_set_i18n("en_US")
gtl_opt_short_language(valid = c("de", "fr"))
}
\seealso{
\code{\link[=gtl_opt_set_i18n]{gtl_opt_set_i18n()}}
}
