% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{gtl_options}
\alias{gtl_options}
\title{Set or get options for my package}
\usage{
gtl_options(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \verb{[key]=[value]} pairs to set options.}
}
\description{
Set or get options for my package
}
\section{Supported options}{

The following options are supported
\describe{
\item{\code{language}}{The default language}
\item{\code{country}}{The default country setting}
\item{\code{theme}}{The default ggplot theme}
\item{\code{mode}}{The default ggplot mode}
\item{\code{opacity}}{The default opacity}
}
}

\examples{
# Retrieve default options
gtl_options("language")
gtl_options("plot_standard_width")

# Change a setting and retrieve it
gtl_options(plot_standard_width = 25)
gtl_options("plot_standard_width")
}
