% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_and_labels.R
\name{gtl_relabel_dash}
\alias{gtl_relabel_dash}
\title{Reformat santoku chop dash labels}
\usage{
gtl_relabel_dash(x)
}
\arguments{
\item{x}{A string vector of labels}
}
\value{
A string vector with clean labels
}
\description{
This function can be used to reformat the levels produced
by the \code{\link[santoku:lbl_dash]{santoku::lbl_dash()}} function.
}
\examples{
data <- seq(1, 20)
data_cut <- santoku::chop(data, breaks = c(5,7,13),
                          labels = santoku::lbl_dash(),
                          extend = TRUE, drop = FALSE)
forcats::fct_relabel(data_cut, gtl_relabel_dash)

}
\seealso{
\code{\link[santoku:lbl_dash]{santoku::lbl_dash()}}
}
