% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{gtl_sci_10_html}
\alias{gtl_sci_10_html}
\title{Format numbers using scientific notation for html}
\usage{
gtl_sci_10_html(x)
}
\arguments{
\item{x}{Number to format}
}
\value{
A formatted expression
}
\description{
Format numbers using scientific notation for html
}
