% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{gtl_swiss_canton_abbr}
\alias{gtl_swiss_canton_abbr}
\title{Get Swiss canton abbreviations}
\usage{
gtl_swiss_canton_abbr(canton_name)
}
\arguments{
\item{canton_name}{A vector of Swiss canton names}
}
\value{
A vector of Swiss canton abbreviations
}
\description{
This is a shortcut for Swiss canton abbreviation based
on the canton name.
}
\examples{
gtl_swiss_canton_abbr(c("Jura", "Genève", "Neuchâtel"))

}
