% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{gtl_swiss_canton_id}
\alias{gtl_swiss_canton_id}
\title{Get Swiss canton FSO ids}
\usage{
gtl_swiss_canton_id(canton_name)
}
\arguments{
\item{canton_name}{A vector of Swiss canton names}
}
\value{
A vector of Swiss canton FSO ids
}
\description{
This is a shortcut for Swiss canton codes based on the
canton name.
}
\examples{
gtl_swiss_canton_id(c("Jura", "Genève", "Neuchâtel"))

}
