% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.R
\name{gtl_translate_enfr}
\alias{gtl_translate_enfr}
\title{Translate function}
\usage{
gtl_translate_enfr(english, french)
}
\arguments{
\item{english}{English string}

\item{french}{French string

A function that returns the english or french argument depending
on the current language setting.}
}
\value{
String depending on package language option
}
\description{
Translate function
}
\examples{
gtl_opt_set_i18n("fr_CH") # Set language to french
gtl_translate_enfr("Hello world!", "Bonjour le monde !")

gtl_opt_set_i18n("en_US") # Set language to english
gtl_translate_enfr("Hello world!", "Bonjour le monde !")
}
\seealso{
\code{\link[=gtl_opt_set_i18n]{gtl_opt_set_i18n()}}
}
