% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.R
\name{gtl_translator}
\alias{gtl_translator}
\title{Creates translation function for named string list}
\usage{
gtl_translator(dictionary)
}
\arguments{
\item{dictionary}{A list of translations}
}
\value{
A function that translates
}
\description{
Creates translation function for named string list
}
\examples{
translations <- list(
  "maison" = "house",
  "chapeau" = "hat",
  "chat" = "cat"
)

translate_to_english <- gtl_translator(translations)
translate_to_english("maison")
translate_to_english("chat")

}
