% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr2poly.R
\name{hr2poly}
\alias{hr2poly}
\title{Aggregate high resolution grid data onto polygons}
\usage{
hr2poly(data, polygon_type = "mun", col_names = NULL, stats = NULL,
crs = 4326, ncores = 1, keep_geometry = FALSE)
}
\arguments{
\item{data}{Space-time dataset in \code{data.frame} format.}

\item{polygon_type}{Level of detail for the administrative boundaries.
Accepted values are \code{"mun"} (\strong{municipalities}),
\code{"prov"} (\strong{provinces}), or \code{"reg"} (\strong{regions}).
Default values is \code{"mun"}.}

\item{col_names}{Named list with the field \strong{coords_x},
\strong{coords_y} and \strong{t}. These parameters specify the names of the
columns representing the x-coordinate, the y-coordinate and time respectively.
If not specified, then default column names \code{"longitude"},
\code{"latitude"} and \code{"time"} are assumed.}

\item{stats}{Named list specifying which statistics to compute for each
variable of interest. Accepted statistics are \code{"min"} (\strong{minimum}),
\code{"mean"} (\strong{mean value}), \code{"median"} (\strong{median value}),
\code{"max"} (\strong{maximum}), and \code{"std"} (\strong{standard deviation}).
If not specified, then all these statistics are computed for all numeric
variables.}

\item{crs}{Coordinate Reference System (CRS) for the input data, given as an
EPSG code. If it differs from \code{4326} (\strong{WGS 84}), then data will
be reprojected accordingly.}

\item{ncores}{Number of CPU cores to use for parallel processing.
Default value is \code{1}, then computation is performed sequentially. Higher
values enable parallel execution to improve performance on large datasets.}

\item{keep_geometry}{Logical. Whether to retain the geometry in the output
dataset. Default is \code{FALSE}.}
}
\value{
A \code{data.frame} containing daily statistics of the variables of
interest, computed for each administrative boundary (e.g., municipality).
}
\description{
Aggregates input high-resolution space-time data onto \strong{polygons} and
computes the user-specified statistics (e.g., mean, median, and standard
deviation) for each space-time variable. This function supports parallel
computing.
}
\examples{
# SEE "demo.Rmd" FOR MORE DETAILS

\dontrun{
res_hr2poly <- hr2poly(
  data = input_data,
  polygon_type = "mun",
  col_names = list(
    "coords_x" = "x",
    "coords_y" = "y",
    "t" = "time"
  ),
  stats = list(
    "AQ_EEA_NO2" = c("min", "mean", "max", "std"),
    "AQ_CAMS_NO2" = c("min", "median", "max", "std"),
    "altitude" = c("mean")
  ),
  crs = 4979,
  ncores = 2,
  keep_geometry = TRUE
)}

}
\seealso{
\url{https://github.com/GRINS-Spoke0-WP2/geotools/blob/develop/demo/demo.Rmd}
}
