% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idw2hr.R
\name{idw2hr}
\alias{idw2hr}
\title{Interpolate data onto a high-resolution grid using IDW}
\usage{
idw2hr(data, crs = 4326, outgrid_params = NULL, col_names = NULL,
interest_vars = NULL, idp = 1, nmax = 4, ncores = 1, restore_NA = FALSE)
}
\arguments{
\item{data}{Space-time dataset in \code{data.frame} or \code{array} format.}

\item{crs}{Coordinate Reference System (CRS) for the input data, given as an
EPSG code. If it differs from \code{4326} (\strong{WGS 84}), then data will
be reprojected accordingly.}

\item{outgrid_params}{Named list with the fields \strong{resolution},
\strong{min_lon}, \strong{max_lon}, \strong{min_lat} and \strong{max_lat}.
These numeric parameters specify the resolution and geographic bounds of the
output high-resolution grid. All values must refer to the CRS \strong{4326}.
If not specified, then the resolution defaults to \code{0.01} and the
geographic bounds are automatically inferred from the input data.}

\item{col_names}{Named list with the field \strong{coords_x},
\strong{coords_y} and \strong{t}. These parameters specify the names of the
columns representing the x-coordinate, the y-coordinate and time respectively.
If not specified, then default column names \code{"longitude"},
\code{"latitude"} and \code{"time"} are assumed.}

\item{interest_vars}{List specifying the names of the variables to be
interpolated using IDW. If not specified, then all the available variables
will be interpolated.}

\item{idp}{Inverse distance power parameter used in IDW interpolation.
Controls how strongly the weighting decreases with distance. Higher values
give more influence to nearby points. Default value is \code{1}.}

\item{nmax}{Number of points to take into account to perform IDW. Default
value is \code{4}.}

\item{ncores}{Number of CPU cores to use for parallel processing.
Default value is \code{1}, then computation is performed sequentially. Higher
values enable parallel execution to improve performance on large datasets. If
\code{ncores} exceeds the number of available cores, the number of available
cores will be used instead.}

\item{restore_NA}{If \code{TRUE}, the function will propagate NA values from
the low-resolution dataset to the high-resolution dataset by setting
corresponding points and their neighbors to NA. If \code{FALSE}, the function
will leave the high-resolution dataset unchanged regarding NA values. Default
is \code{FALSE}.}
}
\value{
A \code{data.frame} resulting from interpolation onto a
high-resolution grid using IDW.
}
\description{
Interpolates input space-time data onto a high-resolution spatial grid using
\strong{IDW} (\strong{Inverse Distance Weighting}). The goal is to assign
(interpolated) observations within the boundaries of smaller municipalities
(which are polygons) as well. This function supports parallel computing.
}
\examples{
# SEE "demo.Rmd" FOR MORE DETAILS

\dontrun{
res_idw2hr <- idw2hr(
  data = input_data,
  crs = 4979,
  outgrid_params = list(
    "resolution" = 0.02,
    "min_lon" = 6,
    "max_lon" = 7,
    "min_lat" = 42,
    "max_lat" = 43
  ),
  col_names = list(
    "coords_x" = "x",
    "coords_y" = "y",
    "t" = "time"
  ),
  interest_vars = list(
    "altitude",
    "AQ_EEA_NO2",
    "AQ_CAMS_NO2"
  ),
  idp = 1,
  ncores = 4
)}

}
\seealso{
\url{https://github.com/GRINS-Spoke0-WP2/geotools/blob/develop/demo/demo.Rmd}
}
