% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_map_facilities.R
\name{leaflet_map_facilities}
\alias{leaflet_map_facilities}
\title{Map facility-specific locations (as pins), or quantities (as circles), with leaflet}
\usage{
leaflet_map_facilities(
  facilities,
  size = NULL,
  color = "orange",
  opacity = 0.333,
  popup = NULL,
  max_size = 100,
  stroke = FALSE,
  popupOptions = popupOptions(),
  ...,
  verbose = getOption("verbose")
)
}
\arguments{
\item{facilities}{(geodata) for example: \code{point_source_facilities(RY(2015))}}

\item{size}{(expression) if supplied, controls size of circles (area, not radius, will be proportional to this)}

\item{color}{(expression) if supplied, controls colors of circles}

\item{max_size}{(numeric) helpful for tweaking the relative sizes of circles}

\item{stroke}{(logical) see \code{\link[leaflet:addCircleMarkers]{leaflet::addCircleMarkers()}}}

\item{...}{further arguments to \code{\link[=leaflet_SFBA]{leaflet_SFBA()}}}

\item{verbose}{(logical)}
}
\description{
Map facility-specific locations (as pins), or quantities (as circles), with leaflet
}
\seealso{
\code{\link[=leaflet_map_SFBA]{leaflet_map_SFBA()}}
}
