% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_raster.R
\name{mask_raster}
\alias{mask_raster}
\title{mask_raster}
\usage{
mask_raster(x, mask, ...)
}
\arguments{
\item{x}{\verb{Raster*} object}

\item{mask}{\code{sf} object}

\item{...}{reserved for future use}
}
\value{
masked \verb{Raster*} object
}
\description{
Thin wrapper around \code{\link[raster:mask]{raster::mask()}}. Works with an \code{sf} object, and automatically reprojects it to match the CRS of \code{x}.
}
