% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge-methods.R
\name{merge_SpatialDataFrame}
\alias{merge_SpatialDataFrame}
\alias{merge_SpatialPolygonsDataFrame}
\alias{merge_SpatialLinesDataFrame}
\title{Merge generic}
\usage{
merge_SpatialDataFrame(x, y, by, ...)

merge_SpatialPolygonsDataFrame(x, y, by, ...)

merge_SpatialLinesDataFrame(x, y, by, ...)
}
\arguments{
\item{x}{a \code{SpatialLinesDataFrame} object}

\item{y}{data.frame}

\item{by}{character}

\item{\dots}{further arguments, as in \link{merge} generally (e.g. \code{sort})}
}
\value{
an object of the same class as \code{x}
}
\description{
Merge generic
}
\section{Functions}{
\itemize{
\item \code{merge_SpatialPolygonsDataFrame}: Merge a SpatialPolygonsDataFrame with a data.frame

\item \code{merge_SpatialLinesDataFrame}: Merge a SpatialPolygonsDataFrame with a data.frame
}}

