% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meters.R
\name{meters}
\alias{meters}
\title{meters}
\usage{
meters(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Create a numeric vector in units of "meters" (distance)
}
\examples{
meters(5)
meters(100) * meters(100)

}
