% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_missing.R
\name{non_missing}
\alias{non_missing}
\title{Return all rows of a Spatial*DataFrame object except those with NA in the given column}
\usage{
non_missing(spdf, variable)
}
\arguments{
\item{spdf}{\code{Spatial*DataFrame} object}

\item{variable}{name of column}
}
\description{
Return all rows of a Spatial*DataFrame object except those with NA in the given column
}
