% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay_raster.R
\name{overlay_raster}
\alias{overlay_raster}
\title{overlay_raster}
\usage{
overlay_raster(
  input_geodata,
  raster_brick,
  fun = "mean",
  progress = TRUE,
  unit = NULL,
  ...,
  verbose = getOption("verbose", default = TRUE)
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
overlay_raster
}
