% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_areas.R
\name{polygon_areas}
\alias{polygon_areas}
\title{Calculate the area of each polygon in a SpatialPolygons* object}
\usage{
polygon_areas(spdf)
}
\arguments{
\item{spdf}{\verb{SpatialPolygons*} object}
}
\description{
Calculate the area of each polygon in a SpatialPolygons* object
}
