% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj4string_LCC.R
\name{proj4string_LCC}
\alias{proj4string_LCC}
\title{proj4string_LCC}
\usage{
proj4string_LCC(
  lon_0 = 0,
  lat_0 = 0,
  lat_1 = 0,
  lat_2 = 0,
  x_0 = 0,
  y_0 = 0,
  units = "m",
  datum = "NAD83",
  ellps = "GRS80",
  ...,
  verbose = getOption("verbose", default = FALSE)
)
}
\arguments{
\item{lon_0}{Longitude of projection center}

\item{lat_0}{Latitude of projection center}

\item{lat_1}{First standard parallel}

\item{lat_2}{Second standard parallel}

\item{x_0}{False easting}

\item{y_0}{False northing}

\item{...}{reserved for future use}

\item{verbose}{logical}
}
\description{
Construct a \code{PROJ4} string for a Lambert Conformal Conic projection.
}
\examples{
proj4string_LCC(lat_1=30, lat_2=60, lat_0=37, lon_0=-120.5, y_0=16000, x_0=220000, a=6.37e6, b=6.37e6, units="km", verbose=TRUE)
}
\references{
\itemize{
\item https://proj.org/operations/projections/lcc.html
}
}
