% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_fips.R
\name{pull_fips}
\alias{pull_fips}
\title{pull_fips}
\usage{
pull_fips(geodata)
}
\arguments{
\item{geodata}{geodata with rownames that are FIPS codes}
}
\description{
Extract state, county, and tract components from rownames.
}
