% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{quick.grid}
\alias{quick.grid}
\title{Quickly build an interpolation grid with the desired cell size}
\usage{
quick.grid(borders.sp, cellsize)
}
\arguments{
\item{borders.sp}{A sp data frame containing the borders of the extension zone. Must be projected and not in geographic CRS}

\item{cellsize}{The size of the cells of the desired interpolation grid expressed in the units of the map}
}
\value{
a sf interpolation grid
}
\details{
https://stackoverflow.com/questions/43436466/create-grid-in-r-for-kriging-in-gstat/43444232
}
\author{
Thomas Goossens
}
