% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_digest.R
\name{raster_digest}
\alias{raster_digest}
\title{raster_digest}
\usage{
raster_digest(rst_layer, algo = "sha1")
}
\arguments{
\item{rst_layer}{(raster) \code{terra} object}

\item{algo}{(character) passed to \code{\link[digest:digest]{digest::digest()}}}
}
\value{
hash, as character, according to \code{algo}
}
\description{
raster_digest
}
