% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_mask.R
\name{raster_mask}
\alias{raster_mask}
\title{raster_mask}
\usage{
raster_mask(rst, mask, ...)
}
\arguments{
\item{rst}{\verb{Raster*} object}

\item{mask}{\code{sf} object}

\item{...}{reserved for future use}
}
\value{
masked \verb{Raster*} object
}
\description{
Thin wrapper around \code{\link[raster:mask]{raster::mask()}}. Works with an \code{sf} object, and automatically reprojects it to match the CRS of \code{x}.
}
