% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_overlay.R
\name{raster_overlay}
\alias{raster_overlay}
\title{raster_overlay}
\usage{
raster_overlay(
  input_geodata,
  rst,
  fun = "mean",
  progress = TRUE,
  unit = NULL,
  ...,
  verbose = getOption("verbose", default = TRUE)
)
}
\arguments{
\item{input_geodata}{\code{sf} or \code{sfc} object}

\item{fun}{name of summarizing function, like "mean"}

\item{progress}{logical}

\item{unit}{character (optional), like "ug/m^3"}

\item{...}{passed to \code{\link[exactextractr:exact_extract]{exactextractr::exact_extract()}}}

\item{verbose}{logical}

\item{raster_brick}{\link[raster:Raster-classes]{raster::Raster}}
}
\value{

}
\description{
raster_overlay
}
