% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_sum.R
\name{raster_sum}
\alias{raster_sum}
\alias{raster_sum__}
\title{raster_sum}
\usage{
raster_sum(
  layer_list,
  weights = NULL,
  na.rm = TRUE,
  ...,
  progress = TRUE,
  cache = TRUE,
  chunk_size = 100,
  verbose = FALSE
)

raster_sum__(
  layer_list,
  weights = NULL,
  progress = NULL,
  chunk_size = 30,
  na.rm = TRUE,
  ...,
  cache = TRUE,
  cache_dir = here::here("Cache", "raster_sum__"),
  verbose = FALSE
)
}
\arguments{
\item{layer_list}{list of something that can be passed to \code{\link[terra:rast]{terra::rast()}}}

\item{weights}{(optional) numeric, of same length as \code{layer_list}}

\item{na.rm}{(logical)}

\item{...}{options for writing files, as in \code{\link[terra:writeRaster]{terra::writeRaster()}}}
}
\description{
raster_sum

raster_sum__
}
\section{Functions}{
\itemize{
\item \code{raster_sum__()}: uncached variant

}}
