% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_kmz.R
\name{read_kmz}
\alias{read_kmz}
\title{read_kmz}
\usage{
read_kmz(kmz_path, kml_file = NULL, crs = WGS84_CRS, ..., verbose = FALSE)
}
\arguments{
\item{kmz_path}{path}

\item{kml_file}{for backwards compatibility}

\item{crs}{passed to \code{reproject()}}

\item{...}{passed to \code{reproject()}; may include \code{coordnames}}

\item{verbose}{display messages}
}
\description{
Read the first KML file contained in a KMZ file.
}
