% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ncdf_raster.R
\name{read_ncdf_raster}
\alias{read_ncdf_raster}
\title{read_ncdf_raster}
\usage{
read_ncdf_raster(path, ..., verbose = getOption("verbose", default = FALSE))
}
\arguments{
\item{path}{character}

\item{...}{reserved for future use}

\item{verbose}{logical}
}
\value{
\code{RasterLayer} object
}
\description{
read_ncdf_raster
}
