% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproject.R
\name{reproject}
\alias{reproject}
\title{Reproject/transform a Spatial* object}
\usage{
reproject(x, new_CRS = EPSG_4326, new_coordnames = c("lng", "lat"))
}
\description{
Reproject/transform a Spatial* object
}
