% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{sp_centroid}
\alias{sp_centroid}
\title{sp_centroid}
\usage{
sp_centroid(spobj, ...)
}
\arguments{
\item{spobj}{\verb{Spatial*} object}

\item{...}{passed to \code{\link[rgeos:topo-unary-gCentroid]{rgeos::gCentroid()}}}
}
\description{
Extract the coordinates, as a simple vector, of the centroid of a Spatial* object.
}
