% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_envelope.R
\name{st_envelope}
\alias{st_envelope}
\title{st_envelope}
\usage{
st_envelope(obj, ...)
}
\arguments{
\item{obj}{simple feature or simple feature set, passed to \code{\link[sf:st_bbox]{sf::st_bbox()}}}

\item{...}{passed tp \link[sf:st_bbox]{sf::st_bbox}}
}
\description{
Return bounding box of \code{obj} as a polygon
}
