% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_extent.R
\name{st_extent}
\alias{st_extent}
\title{st_extent}
\usage{
st_extent(x, ...)
}
\arguments{
\item{x}{\code{sf} or \code{sfc} object}

\item{...}{passed to \code{\link[raster:extent]{raster::extent()}}}
}
\value{
\code{raster::Extent} object
}
\description{
st_extent
}
