% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_filter_spatial.R
\name{st_filter_spatial}
\alias{st_filter_spatial}
\title{Filter one set of geometries based on another}
\usage{
st_filter_spatial(x, y, datum = st_crs(32610), fun = st_intersection,
  view = FALSE, verbose = getOption("verbose"))
}
\arguments{
\item{x}{geometries to filter}

\item{y}{geometries to filter by}

\item{datum}{intermediate datum for projection}

\item{view}{logical; show results in Rstudio viewer?}

\item{verbose}{logical}
}
\description{
Filter one set of geometries based on another
}
