% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_filter_valid.R
\name{st_filter_valid}
\alias{st_filter_valid}
\title{Discard invalid geometries}
\usage{
st_filter_valid(geodata, min_dimension = 1, view = FALSE,
  verbose = TRUE)
}
\arguments{
\item{geodata}{object of type \code{sf}}

\item{min_dimension}{drop geometries with dimension lower than this}

\item{view}{logical; show results in Rstudio viewer?}

\item{verbose}{logical}
}
\description{
Discard invalid geometries
}
\note{
for 2D polygons, set \code{min_dimension} equal to 1
}
