% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_overlay.R
\docType{methods}
\name{st_overlay.geodata.raster}
\alias{st_overlay.geodata.raster}
\alias{st_overlay}
\alias{st_overlay,sf,RasterLayer-method}
\alias{st_overlay,sf,RasterStack-method}
\title{Overlay spatial data onto some kind of (other) spatial data.}
\usage{
st_overlay.geodata.raster(geodata, object, ...)

st_overlay(geodata, object, ...)

\S4method{st_overlay}{sf,RasterLayer}(geodata, object, ...)

\S4method{st_overlay}{sf,RasterStack}(geodata, object, ...)
}
\arguments{
\item{geodata}{for now, should be polygons in the form of an \code{sf} object}

\item{object}{for now, should be a \code{RasterLayer} or \code{RasterStack}}

\item{...}{further arguments (reserved for future use)}
}
\description{
WARNING: Experimental work in progress! Developed to overlay a RasterStack onto polygons. YMMV.
}
