% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{st_repair}
\alias{st_repair}
\title{st_repair}
\usage{
st_repair(geodata, datum = st_crs(32610), view = FALSE, verbose = TRUE)
}
\arguments{
\item{geodata}{\code{sf} object}

\item{datum}{passed to \code{sf::st_transform()}}

\item{view}{(optional) show result using \code{mapview()}}

\item{verbose}{display messages}
}
\description{
Attempt to repair problems in geodata.
}
