% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_sample_points.R
\name{st_sample_points}
\alias{st_sample_points}
\title{st_sample_points}
\usage{
st_sample_points(geodata, size, verbose = getOption("verbose"))
}
\arguments{
\item{geodata}{(sf) most likely polygons or polylines (inside or along which to sample)}

\item{size}{(int) number of points to sample (in or along each feature)}

\item{verbose}{display messages}
}
\value{
\code{sfc} of POINTS, joined back to the original \code{geodata} (i.e., an overlay)
}
\description{
Sample and overlay a set of points in polygons (or along polylines).
}
