% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{static.ggmap}
\alias{static.ggmap}
\title{Build a static map displaying predictions and their related error}
\usage{
static.ggmap(gridded.data.df, boundaries.sf, layer.error.bool,
  legend.error.bool, pretty_breaks.bool, title.chr, target.chr)
}
\arguments{
\item{gridded.data.df}{A data frame obtained from a SpatialGridDataFrame containing data}

\item{boundaries.sf}{A sf containing data from Wallonia boundaries}

\item{layer.error.bool}{A boolean specifying if you want to display the layer with error}

\item{legend.error.bool}{A boolean specifying if you want to display the legend of the error layer}

\item{pretty_breaks.bool}{A boolean specifying the type of legend you want. TRUE for pretty breaks, FALSE for quantile scale}

\item{title.chr}{A character specifying the title you want for your map}

\item{target.chr}{A character specifying the predicted parameter. One of "tsa", "hra" or "hct"}
}
\value{
a ggplot map object
}
\description{
Build a static map displaying predictions and their related error
}
\author{
Loïc Davadan <- ldavadan.github.io
}
