% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.R
\name{translator}
\alias{translator}
\title{Creates translation function for named string list}
\usage{
translator(dictionary)
}
\arguments{
\item{dictionary}{A list of translations}
}
\value{
A function that translates
}
\description{
\code{translator()} was renamed to \code{gtl_translator()} to create a more
consistent API.
}
\keyword{internal}
