% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmute_raster.R
\name{transmute_raster}
\alias{transmute_raster}
\title{transmute_raster}
\usage{
transmute_raster(raster_object, f, verbose = getOption("verbose"))
}
\arguments{
\item{f}{formula}

\item{verbose}{}
}
\value{

}
\description{
transmute_raster
}
