% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzip_url.R
\name{unzip_url}
\alias{unzip_url}
\title{To download and extract a zipped file (that may not have .zip at the end of the URL)}
\usage{
unzip_url(url, destfile, exdir, quiet = FALSE)
}
\arguments{
\item{url}{character}

\item{destfile}{name for local copy of (downloaded) file}

\item{exdir}{directory to hold expanded contents}

\item{quiet}{(optional)}
}
\description{
To download and extract a zipped file (that may not have .zip at the end of the URL)
}
