% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_xy.R
\docType{methods}
\name{with_xy}
\alias{with_xy}
\alias{with_xy,SpatialPoints-method}
\alias{with_xy,SpatialPoints,ANY-method}
\alias{with_xy,SpatialPolygons-method}
\alias{with_xy,SpatialPolygons,ANY-method}
\alias{with_xy,SpatialPixels-method}
\alias{with_xy,SpatialPixels,ANY-method}
\alias{with_xy,SpatialPointsDataFrame-method}
\alias{with_xy,SpatialPolygonsDataFrame-method}
\alias{with_xy,SpatialPolygonsDataFrame,ANY-method}
\alias{with_xy,SpatialPixelsDataFrame-method}
\alias{with_xy,SpatialPixelsDataFrame,ANY-method}
\title{S4 methods for with_xy}
\usage{
with_xy(spobj)

\S4method{with_xy}{SpatialPoints}(spobj)

\S4method{with_xy}{SpatialPolygons}(spobj)

\S4method{with_xy}{SpatialPixels}(spobj)

\S4method{with_xy}{SpatialPointsDataFrame}(spobj)

\S4method{with_xy}{SpatialPolygonsDataFrame}(spobj)

\S4method{with_xy}{SpatialPixelsDataFrame}(spobj)
}
\arguments{
\item{spobj}{\verb{Spatial*} object}
}
\description{
S4 methods for with_xy
}
