% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_xy.R
\name{with_xy.default}
\alias{with_xy.default}
\title{Return a copy of the Spatial* object with coordnames set to
x and y, and a copy of the coordinates stored in its data.frame}
\usage{
with_xy.default(spobj)
}
\description{
Return a copy of the Spatial* object with coordnames set to
x and y, and a copy of the coordinates stored in its data.frame
}
