% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_geojson.R
\name{write_geojson}
\alias{write_geojson}
\title{write_geojson}
\usage{
write_geojson(
  object,
  dsn = NULL,
  layer = NULL,
  ...,
  crs = 4326,
  overwrite = TRUE,
  pretty = TRUE,
  digits = 6,
  append = FALSE,
  verbose = getOption("verbose", default = FALSE)
)
}
\arguments{
\item{dsn}{output path; defaults to the current working directory}

\item{layer}{name of layer to write}

\item{...}{passed to \code{\link[sf:st_write]{sf::st_write()}}}

\item{overwrite}{(logical)}

\item{verbose}{display messages}

\item{spobj}{geodata}
}
\description{
Write a layer to a GeoJSON file.
}
