% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_GeoTIFF.R
\name{write_GeoTIFF}
\alias{write_GeoTIFF}
\title{write_GeoTIFF}
\usage{
write_GeoTIFF(raster_obj, ..., suffix = "names", overwrite = TRUE)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Thin wrapper around \code{\link[raster:writeRaster]{raster::writeRaster()}}
}
