% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_raster.R
\name{write_raster}
\alias{write_raster}
\title{write_raster}
\usage{
write_raster(raster_obj, path, ..., format = "raster", overwrite = TRUE)
}
\arguments{
\item{raster_obj}{\href{raster::Raster-class}{Raster*} object}

\item{path}{character}

\item{...}{passed to \code{\link[raster:writeRaster]{raster::writeRaster()}}}

\item{format}{character (defaults to "raster"); see \code{\link[raster:writeRaster]{raster::writeRaster()}}}

\item{overwrite}{logical}
}
\value{
result of \code{\link[raster:writeRaster]{raster::writeRaster()}}, most likely a handle to an on-disk raster
}
\description{
write_raster
}
