% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countrycode.R
\name{admincode}
\alias{admincode}
\title{Provide custom countrycode function}
\usage{
admincode(..., country = c("Switzerland", "Russia"))
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[countrycode:countrycode]{countrycode::countrycode}}
  \describe{
    \item{\code{sourcevar}}{Vector which contains the codes or country names to be
converted (character or factor)}
    \item{\code{origin}}{Coding scheme of origin (string such as "iso3c" enclosed in
quotes ""): type "?codelist" for a list of available codes.}
    \item{\code{destination}}{Coding scheme of destination (string such as "iso3c"
enclosed in quotes ""): type `?codelist` for a list of
available codes.}
    \item{\code{warn}}{Prints unique elements from sourcevar for which no match was found}
    \item{\code{nomatch}}{When countrycode fails to find a match for the code of
origin, it fills-in the destination vector with nomatch. The default
behavior is to fill non-matching codes with NA. If nomatch = NULL,
countrycode tries to use the origin vector to fill-in missing values in the
destination vector. nomatch must be either NULL, of length 1, or of the same
length as sourcevar.}
    \item{\code{custom_dict}}{A data frame which supplies a new dictionary to replace
the built-in country code dictionary. Each column contains a different code
and must include no duplicates. The data frame format should resemble
`countrycode::codelist`.  Warning: when `custom_dict` is used, no sanity
checks are conducted.}
    \item{\code{custom_match}}{A named vector which supplies custom origin and
destination matches that will supercede any matching default result. The name
of each element will be used as the origin code, and the value of each
element will be used as the destination code.}
    \item{\code{origin_regex}}{Logical: When using a custom dictionary, if TRUE then the
origin codes will be matched as regex, if FALSE they will be matched exactly.
When using the default dictionary (dictionary = NULL), origin_regex will be ignored.}
  }}

\item{country}{Select the custom dictionary, currently one of \emph{switzerland} or \emph{russia}}
}
\value{
A vector with the desired output
}
\description{
Provide custom countrycode function
}
