% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_spatial.R
\name{as_spatial}
\alias{as_spatial}
\title{Create a SpatialPointsDataFrame from tabular data}
\usage{
as_spatial(input_data, coord_vars, crs = CRS(NA_character_), na.rm = TRUE, ...)
}
\description{
Create a SpatialPointsDataFrame from tabular data
}
