% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_raster.R
\name{cut_raster}
\alias{cut_raster}
\title{cut_raster}
\usage{
cut_raster(
  raster_layer,
  breaks,
  include.lowest = TRUE,
  ...,
  na.rm = TRUE,
  verbose = getOption("verbose", default = FALSE)
)
}
\arguments{
\item{raster_layer}{}

\item{breaks}{}

\item{include.lowest}{}

\item{...}{}
}
\value{

}
\description{
cut_raster
}
