% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotools-package.R
\docType{data}
\name{UTM10_NAD83}
\alias{UTM10_NAD83}
\alias{UTM10_NAD27}
\title{UTM10 NAD83 (EPSG:26910)}
\format{
an \code{sp::CRS} object

an \code{sp::CRS} object
}
\usage{
UTM10_NAD83

UTM10_NAD27
}
\description{
Universal Transverse Mercator (UTM), Zone 10, North American Datum (NAD) 1983

Universal Transverse Mercator (UTM), Zone 10, North American Datum (NAD) 1927
}
\keyword{datasets}
