% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{filter_.SpatialDataFrame}
\alias{filter_.SpatialDataFrame}
\title{Filter (subset) a Spatial*DataFrame object}
\usage{
\method{filter_}{SpatialDataFrame}(.data, ..., .dots)
}
\arguments{
\item{.data}{a \code{Spatial*DataFrame}}

\item{...}{see \link{subset}}

\item{.dots}{}
}
\value{
a subset of the original \code{Spatial*DataFrame}
}
\description{
Filter (subset) a Spatial*DataFrame object
}
\seealso{
subset
}
