% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_spatial.R
\name{filter_spatial}
\alias{filter_spatial}
\title{filter_spatial}
\usage{
filter_spatial(x, y, FUN = st_intersects, ..., verbose = getOption("verbose"))
}
\arguments{
\item{x}{object to be filtered}

\item{y}{object to filter with}

\item{FUN}{(function) a spatial predicate, for example: \code{\link[sf:geos_binary_pred]{sf::st_intersects()}}}

\item{...}{further arguments to \code{\link[sf:st_join]{sf::st_filter()}}}

\item{verbose}{(logical)}
}
\value{
(typically) a subset of \code{x}
}
\description{
A slightly more convenient version of \code{\link[sf:st_join]{sf::st_filter()}}. It does the
job of reprojecting the second object to the same CRS as the first,
if that's needed.
}
