% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_spatial.R
\name{filter_spatial}
\alias{filter_spatial}
\title{Apply a spatial filter}
\usage{
filter_spatial(x, y, FUN = NULL, ..., verbose = getOption("verbose"))
}
\arguments{
\item{FUN}{(function) a spatial predicate, for example: \link[rgeos:gIntersects]{gIntersects}}

\item{...}{further arguments to \code{FUN}}

\item{verbose}{(logical)}

\item{spobj1}{(\link[sp:Spatial]{Spatial} object) to be filtered}

\item{spobj2}{(\link[sp:Spatial]{Spatial} object) to filter with}
}
\value{
(typically) a subset of \code{spobj1}
}
\description{
Apply a spatial filter
}
