% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force_crs.R
\name{force_crs}
\alias{force_crs}
\title{Force replacement of CRS while leaving coordinate values unchanged}
\usage{
force_crs(geodata, crs, ..., verbose = getOption("verbose"))
}
\arguments{
\item{geodata}{(sf or sfc)}

\item{...}{further arguments to \code{\link[sf:st_as_sf]{sf::st_as_sf()}}}

\item{verbose}{(logical)}

\item{coords}{(character) names of the (two) columns containing coordinates}
}
\description{
Force replacement of CRS while leaving coordinate values unchanged
}
