% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gContainsOrOverlaps.R
\name{gContainsOrOverlaps}
\alias{gContainsOrOverlaps}
\title{gContainsOrOverlaps}
\usage{
gContainsOrOverlaps(geom2, geom1, byid = TRUE)
}
\description{
Default basis for \code{gFilter} is \code{gContainsOrOverlaps}
}
\note{
Unlike most \code{rgeos} functions, this returns row.names !!
}
