% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotools-package.R
\docType{package}
\name{geotools-package}
\alias{geotools}
\alias{geotools-package}
\title{geotools: Manipulate and Filter Geospatial Data}
\description{
Provides simplified, user-friendly ways to work with geodata.
}
\keyword{internal}
