% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotools-package.R
\docType{package}
\name{geotools-package}
\alias{geotools}
\alias{geotools-package}
\title{geotools: Provide Support for Geography Packages}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Support functions for the geographer package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://geotools.nenuial.org}
}

}
\author{
\strong{Maintainer}: Pascal Burkhard \email{pascal.burkhard@gmail.com} (\href{https://orcid.org/0009-0001-5504-5084}{ORCID})

}
\keyword{internal}
