% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{gtl_chk_rnoaa_api_key}
\alias{gtl_chk_rnoaa_api_key}
\title{Check NOAA API key}
\usage{
gtl_chk_rnoaa_api_key()
}
\value{
Returns an error if the NOAA API key is not properly set
}
\description{
Function to check if the NOAA API key is correctly set.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Not run: fails if the NOAA API key isn't set

gtl_chk_rnoaa_api_key()
\dontshow{\}) # examplesIf}
}
