% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{gtl_country_list_noaa}
\alias{gtl_country_list_noaa}
\title{NOAA country list}
\usage{
gtl_country_list_noaa()
}
\description{
This function returns a list of World Bank countries
for a choice selector in a Shiny app.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Not run: need the NCDC city database
gtl_country_list_noaa()
\dontshow{\}) # examplesIf}
}
