% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{gtl_crs_proj}
\alias{gtl_crs_proj}
\title{Get a defined CRS definition for a given code}
\usage{
gtl_crs_proj(code)
}
\arguments{
\item{code}{A string with the CRS identifier}
}
\value{
A CRS object
}
\description{
A function that returns the corresponding CRS for different
codes.
}
\details{
The current list of available codes is:
\itemize{
\item eqearth
\item equirec
\item gallpeters
\item goode
\item hobodyer
\item mercator
\item robinson
\item wintri
}
}
\examples{
gtl_crs_proj("eqearth")

gtl_crs_proj("robinson")

}
