% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{gtl_crs_regional}
\alias{gtl_crs_regional}
\title{Get a defined CRS for a given country}
\usage{
gtl_crs_regional(country)
}
\arguments{
\item{country}{A string with the country name}
}
\value{
A crs object
}
\description{
Get a defined CRS for a given country
}
