% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_and_labels.R
\name{gtl_full_date}
\alias{gtl_full_date}
\title{Format full date}
\usage{
gtl_full_date(date)
}
\arguments{
\item{date}{A date in YYYY-MM-DD format}
}
\value{
A formated \verb{day month year} (\%d \%B \%Y) date
}
\description{
This function formats a date in a long format (see return value). The format
takes the current language setting into account (see \code{\link[=gtl_opt_set_i18n]{gtl_opt_set_i18n()}}).
}
\examples{
# Set en english locale
gtl_opt_set_i18n("en_US")
gtl_full_date("2020-02-20")

# Set a french locale
gtl_opt_set_i18n("fr_FR")
gtl_full_date("2020-02-20")

}
\seealso{
\code{\link[=gtl_opt_set_i18n]{gtl_opt_set_i18n()}}
}
