% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.R
\name{gtl_koppen_code}
\alias{gtl_koppen_code}
\title{Determine the Köppen climate based on temperature,
precipitation and the latitude.}
\usage{
gtl_koppen_code(temp, prec, lat)
}
\arguments{
\item{temp}{A vector with temperatures for each month}

\item{prec}{A vector with precipitations for each month}

\item{lat}{The latitude}
}
\value{
A string with the Köppen climate determination
}
\description{
Determine the Köppen climate based on temperature,
precipitation and the latitude.
}
