% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{gtl_opt_long_language}
\alias{gtl_opt_long_language}
\title{Return long language name}
\usage{
gtl_opt_long_language()
}
\value{
A string with the language name (lowercase)
}
\description{
Return long language name
}
