% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{gtl_opt_set_i18n}
\alias{gtl_opt_set_i18n}
\title{Set language and country settings}
\usage{
gtl_opt_set_i18n(code)
}
\arguments{
\item{code}{A valid i18n code (example: fr_CH)}
}
\description{
Set language and country settings
}
