% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{gtl_opt_short_language}
\alias{gtl_opt_short_language}
\title{Return short language code}
\usage{
gtl_opt_short_language(valid = c("en", "fr"))
}
\arguments{
\item{valid}{A vector with valid language options}
}
\value{
A string
}
\description{
Return short language code
}
