% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{gtl_options}
\alias{gtl_options}
\title{Set or get options for my package}
\usage{
gtl_options(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\description{
Set or get options for my package
}
\section{Supported options}{

The following options are supported
\itemize{
 \item{\code{language}}{The default language}
 \item{\code{country}}{The default country setting}
 \item{\code{theme}}{The default ggplot theme}
 \item{\code{mode}}{The default ggplot mode}
 \item{\code{opacity}}{The default opacity}
}
}

