% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{leafletize}
\alias{leafletize}
\title{Build a leaflet map displaying predictions and their related error}
\usage{
leafletize(data.sf, se.bool = TRUE)
}
\arguments{
\item{data.sf}{A sf data frame containing the spztilized data}

\item{se.bool}{logial specifying wether to display or not the se}
}
\value{
a leaflet map object
}
\description{
Build a leaflet map displaying predictions and their related error
}
\author{
Thomas Goossens
}
