% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay_raster.R
\name{overlay_raster}
\alias{overlay_raster}
\title{overlay_raster}
\usage{
overlay_raster(
  input_geodata,
  raster_brick,
  fun = "mean",
  progress = TRUE,
  unit = NULL,
  ...,
  verbose = getOption("verbose", default = TRUE)
)
}
\arguments{
\item{input_geodata}{\code{sf} or \code{sfc} object}

\item{raster_brick}{\link[raster:Raster-classes]{raster::Raster}}

\item{fun}{name of summarizing function, like "mean"}

\item{progress}{logical}

\item{unit}{character (optional), like "ug/m^3"}

\item{...}{passed to \code{\link[exactextractr:exact_extract]{exactextractr::exact_extract()}}}

\item{verbose}{logical}
}
\value{

}
\description{
overlay_raster
}
