% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{polygon_areas}
\alias{polygon_areas}
\title{Merge generic}
\usage{
polygon_areas(spdf)
}
\arguments{
\item{spdf}{\verb{SpatialPolygons*} object}

\item{y}{data.frame}

\item{by}{character}

\item{\dots}{further arguments, as in \link{merge} generally (e.g. \code{sort})

Calculate the area of each polygon in a SpatialPolygons* object}
}
\description{
Merge generic
}
