% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ncdf_raster.R
\name{read_ncdf_raster}
\alias{read_ncdf_raster}
\title{read_ncdf_raster}
\usage{
read_ncdf_raster(
  path,
  ...,
  layer_name = NULL,
  crs = NULL,
  verbose = getOption("verbose", default = FALSE)
)
}
\arguments{
\item{path}{character}

\item{...}{reserved for future use}

\item{verbose}{logical}
}
\value{
\code{SpatRaster} object
}
\description{
read_ncdf_raster
}
