% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_envelope.R
\name{st_envelope}
\alias{st_envelope}
\title{st_envelope}
\usage{
st_envelope(x, ..., crs = NULL)
}
\arguments{
\item{x}{either a spatial object, or a vector of four coordinates \code{c(xmin, xmax, ymin, ymax)}}

\item{...}{passed to \link[sf:st_bbox]{sf::st_bbox} or \code{\link[sf:st_as_sf]{sf::st_as_sf()}}}
}
\description{
Construct a rectangular polygon enveloping \code{x}
}
