% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.R
\name{translate_enfr}
\alias{translate_enfr}
\title{Translate function}
\usage{
translate_enfr(english, french)
}
\arguments{
\item{english}{English string}

\item{french}{French string}
}
\value{
String depending on package language option
}
\description{
\code{translate_enfr()} was renamed to \code{gtl_translate_enfr()} to create a more
consistent API.
}
\keyword{internal}
