% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmute_raster.R
\name{transmute_raster}
\alias{transmute_raster}
\title{transmute_raster}
\usage{
transmute_raster(raster_object, f, verbose = getOption("verbose"))
}
\arguments{
\item{raster_object}{multi-layer \href{raster::Raster-class}{Raster*} object}

\item{f}{formula}

\item{verbose}{logical}

\item{...}{ignored}
}
\description{
transmute_raster
}
