% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_geojson.R
\name{write_geojson}
\alias{write_geojson}
\title{write_geojson}
\usage{
write_geojson(
  spobj,
  dsn = getwd(),
  layer = deparse(substitute(spobj)),
  pretty = TRUE,
  digits = 6,
  ...,
  overwrite_layer = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{spobj}{geodata}

\item{dsn}{output path; defaults to the current working directory}

\item{layer}{name of layer to write}

\item{pretty}{passed to \code{jsonlite::toJSON()}}

\item{digits}{passed to \code{jsonlite::toJSON()}}

\item{...}{passed to \code{rgdal::writeOGR()}}

\item{overwrite_layer}{passed to \code{rgdal::writeOGR()}}

\item{verbose}{display messages}
}
\description{
Write a layer to a GeoJSON file.
}
