% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_GeoTIFF.R
\name{write_GeoTIFF}
\alias{write_GeoTIFF}
\title{write_GeoTIFF}
\usage{
write_GeoTIFF(raster_obj, ..., suffix = "names", overwrite = TRUE)
}
\arguments{
\item{raster_obj}{\href{raster::Raster-class}{Raster*} object}

\item{...}{passed to \code{\link[=write_raster]{write_raster()}}}
}
\value{
result of \code{write_raster()}
}
\description{
Thin wrapper around \code{\link[=write_raster]{write_raster()}}
}
\seealso{
\itemize{
\item \code{\link[=write_raster]{write_raster()}}
}
}
